requirements.txt
faiss-cpu
numpy
torch
transformers
sentence-transformers
rank-bm25
langchain
pymupdf
